LPS-200037 move methods to only usage on GCloudNaturalLanguageDocumentAssetAutoTagProvider, delete tagger class and rename test class
